test: Rename testsuite/clean_new_layout.rs to testsuite/clean.rs
